requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.12.57
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.5
sqlparse==0.4.3